requirements.txt
Flask==2.2.2
werkzeug==2.2.2
Flask-SQLAlchemy==3.0.2
flask-mysql==1.5.2
flask-restx==1.0.3
PyJWT==2.6.0
datetime==4.8
pysha3==1.0.2
Flask-Cors===3.0.10
cryptography===39.0.0
